% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myhdi.R
\name{myhdi}
\alias{myhdi}
\title{Highest Density Interval}
\usage{
myhdi(funame, credwidth = 0.95, tol = 1e-08, ...)
}
\arguments{
\item{funame}{q R function}

\item{credwidth}{internval width desired}

\item{tol}{this is the tolerance for the q function}

\item{...}{parameters for the q function}
}
\value{
list of two components, L and U
}
\description{
Highest Density Interval
}
\examples{
# Example with Beta distribution
myhdi(funame = qbeta, credwidth = 0.95, tol = 1e-8, shape1 = 5, shape2 = 7)

# Example with Gamma distribution
myhdi(funame = qgamma, credwidth = 0.95, tol = 1e-8, rate = 2, shape = 4)

}
